requirements.txt
absl-py==2.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
google-auth==2.27.0
google-auth-oauthlib==1.0.0
grpcio==1.60.0
gymnasium==0.29.1
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.7.4
mpmath==1.3.0
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
packaging==23.2
pandas==2.0.3
pillow==10.2.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stable-baselines3==2.2.1
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
torch==2.2.0
tornado==4.5.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
Werkzeug==3.0.1
zipp==3.17.0
